DAO da compra implementado.

+++ finalCore/src/finalCore/dao/CupomDAO.java
@@ -56,7 +56,7 @@ public class CupomDAO extends AbstractJdbcDAO {
 
 				c.setId(rs.getInt("id_cupom"));
 				c.setCodigo(rs.getString("codigo"));
-				c.setDesconto(rs.getInt("desconto"));
+				c.setDesconto(rs.getDouble("desconto"));
 				cupons.add(c);
 			}
 			return cupons;
